Attempting to fix flakiness with PhoenixTracingEndToEndTest
Making a single sink w/ listeners doesn't seem to solve everything. Somehow the connection
is getting closed early and not writing (maybe).

Maybe an open/closed latch on the writer to prevent writes outside the test scope - there
are a bunch of things that happen in the usual course of the tests that might also write
to the metrics system...